refactor(run): split manager.go into purpose-named files
manager.go had grown to ~4,720 lines, with the Create setup pipeline alone
spanning ~2,500. Carve the Manager methods out of the single god-file into
six purpose-named files, all still package run:

  manager.go              core: types, constructors, runtime helpers,
                          accessors, Get/List/Close
  manager_create.go       Create + its host-side setup helpers
  manager_lifecycle.go    Start, StartAttached, Stop, Wait, Destroy
  manager_monitor.go      container-exit and proxy-health monitors
  manager_cleanup.go      log capture, provider stop hooks, resource cleanup
  manager_persistence.go  reconstructing run state from on-disk metadata
  manager_exec.go         exec, clipboard, TTY resize, log access

This is a pure mechanical move: no logic, signatures, or behavior change.
Every top-level declaration is preserved verbatim (verified: the set of
production decls is byte-identical before and after). Each new file gets a
one-line header comment describing what it holds.

First step of the manager.go decomposition; phase extractions (e.g. folding
the credential store opened 7x in Create down to once) follow in later PRs.
